branches/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-storages==1.11.1
django-filter==2.4.0
boto3==1.17.22
Pillow==2.2.1
psycopg2-binary==2.8.5
